+++ phaser/src/battle/logic.ts
@@ -30,188 +30,25 @@ export function randIntBetween(nonce: Uint8Array, index: number, min: number, ma
 }
 
 /**
- * Runs a simulation of the combat round logic found in the combat_round() circuit.
- * 
+ * Combat round logic that accepts specific player targets
  * @param battle_id Battle to be simulated. This is looked up int gameState so the battle must have been created first
  * @param gameState Current game's state. This is both input and output. Modified during execution.
+ * @param playerTargets Player-selected targets for each spirit. This is an array of numbers, where each number corresponds to the enemy index (0, 1, 2) that the spirit will target.
  * @param uiHooks Optional callbacks that can hook into UI animations when calling this for frontend simulation.
  * @returns Rewards from the battle if it is completed (all enemies died or player died). or undefined if it remains in progress
  */
-export function combat_round_logic(battle_id: bigint, gameState: Game2DerivedState, uiHooks?: CombatCallbacks): Promise<BattleRewards | undefined> {
+export function combat_round_logic(battle_id: bigint, gameState: Game2DerivedState, playerTargets?: number[], uiHooks?: CombatCallbacks): Promise<BattleRewards | undefined> {
     return new Promise(async (resolve) => {
-        logger.combat.debug(`combat_round_logc(${safeJSONString(gameState)}, ${uiHooks == undefined})`);
+        logger.combat.debug(`combat_round_logic(${playerTargets}, ${uiHooks == undefined})`);
         const battleConfig = gameState.activeBattleConfigs.get(battle_id)!;
         const battleState = gameState.activeBattleStates.get(battle_id)!;
-        const rng = pureCircuits.fakeTempRng(battleState, battleConfig);
-        if (uiHooks != undefined) {
-            gameState.ui = true;
-        } else {
-            gameState.circuit = true;
+
+        if (playerTargets == undefined) {
+            // If no playerTargets are provided, default to targeting first enemy
+            logger.combat.debug(`No player targets provided, defaulting to first enemy`);
+            playerTargets = new Array(Number(battleConfig.enemy_count)).fill(0).map((_, i) => i);
         }
 
-        const abilityIds = battleState.deck_indices.map((i) => battleConfig.loadout.abilities[Number(i)]);
-        const abilities = abilityIds.map((id) => gameState.allAbilities.get(id)!)
-
-        let player_block = BigInt(0);
-        const enemy_count = Number(battleConfig.enemy_count);
-        let player_damage = new Array(enemy_count).fill(BigInt(0));
-        let enemy_damage = BigInt(0);
-
-        // to be able to early-abort we have this in a lambda
-        const handleEndOfRound = () => {
-            logger.combat.debug(`${uiHooks == undefined} handleEndOfRound.player_damage = ${player_damage}`);
-            uiHooks?.onEndOfRound();
-            if (enemy_damage > player_block) {
-                battleState.player_hp -= enemy_damage - player_block;
-            }
-            if (player_damage[0] > battleConfig.stats[0].block) {
-                battleState.enemy_hp_0 = BigInt(Math.max(0, Number(battleState.enemy_hp_0 + battleConfig.stats[0].block - player_damage[0])));
-            }
-            if (player_damage[1] > battleConfig.stats[1].block) {
-                battleState.enemy_hp_1 = BigInt(Math.max(0, Number(battleState.enemy_hp_1 + battleConfig.stats[1].block - player_damage[1])));
-            }
-            if (player_damage[2] > battleConfig.stats[2].block) {
-                battleState.enemy_hp_2 = BigInt(Math.max(0, Number(battleState.enemy_hp_2 + battleConfig.stats[2].block - player_damage[2])));
-            }
-            logger.combat.debug(`Player HP ${battleState.player_hp} | Enemy HP: ${battleState.enemy_hp_0} / ${battleState.enemy_hp_1} / ${battleState.enemy_hp_2}`);
-            if (battleState.player_hp <= 0) {
-                logger.combat.info(`YOU DIED`);
-                resolve({ alive: false, gold: BigInt(0), ability: { is_some: false, value: BigInt(0) } });
-            }
-            else if (battleState.enemy_hp_0 <= 0 && (battleState.enemy_hp_1 <= 0 || battleConfig.enemy_count < 2) && (battleState.enemy_hp_2 <= 0 || battleConfig.enemy_count < 3)) {
-                logger.combat.info(`YOU WON`);
-                // TODO how to determine rewards?
-                let abilityReward = { is_some: false, value: BigInt(0) };
-                for (let i = 0; i < battleConfig.enemy_count; ++i) {
-                    if (battleConfig.stats[i].boss_type != BOSS_TYPE.normal) {
-                        const ability = generateRandomAbility(BigInt(2));
-                        const abilityId = pureCircuits.derive_ability_id(ability);
-                        // TODO: this really shouldn't be here, should it? but if we don't do that we need to return the entire ability in the contract
-                        // if we don't return it, we need to match the logic here with the contract
-                        gameState.allAbilities.set(abilityId, ability);
-                        abilityReward.is_some = true;
-                        abilityReward.value = abilityId;
-                        break;
-                    }
-                }
-                resolve({ alive: true, gold: BigInt(randIntBetween(rng, 1000, 50, 200)), ability: abilityReward });
-            } else {
-                logger.combat.info(`CONTINUE BATTLE`);
-                resolve(undefined);
-            }
-            logger.combat.debug(`end state[${uiHooks == undefined}]: ${safeJSONString(gameState)}`);
-        }
-
-        await uiHooks?.onDrawAbilities(abilities);
-
-        // enemy block
-        let enemy_hp = [battleState.enemy_hp_0, battleState.enemy_hp_1, battleState.enemy_hp_2];
-        for (let i = 0; i < battleConfig.enemy_count; ++i) {
-            if (enemy_hp[i] > 0) {
-                // do not change vars for block since it's directly checked during player against enemy damage code
-                const block = Number(battleConfig.stats[i].block);
-                logger.combat.debug(`${uiHooks == undefined}, yes block ${i}`);
-                if (block != 0) {
-                    await uiHooks?.onEnemyBlock(i, block);
-                }
-            } else {
-                logger.combat.debug(`${uiHooks == undefined} skipping block ${i}`);
-            }
-        }
-
-        const aliveTargets = new Array(Number(battleConfig.enemy_count))
-            .fill(0)
-            .map((_, i) => i)
-            .filter((i) => enemy_hp[i] > BigInt(0));
-
-        // player abilities (all)
-        const resolveEffect = async (effect: { is_some: boolean, value: Effect }, source: number, target: number) => {
-            if (effect.is_some) {
-                const targets = effect.value.is_aoe ? aliveTargets : [target];
-                switch (effect.value.effect_type) {
-                    case EFFECT_TYPE.attack_fire:
-                    case EFFECT_TYPE.attack_ice:
-                    case EFFECT_TYPE.attack_phys:
-                        const amounts = targets.map((enemy) => {
-                            const dmg = pureCircuits.effect_damage(effect.value, battleConfig.stats[enemy]);
-                            player_damage[enemy] += dmg;
-                            logger.combat.debug(`[${uiHooks == undefined}] player_damage[${enemy}] = ${player_damage[enemy]} // took ${dmg} damage`);
-                            return Number(dmg)
-                        });
-                        await uiHooks?.onPlayerEffect(source, targets, effect.value.effect_type, amounts);
-                        break;
-                    case EFFECT_TYPE.block:
-                        await uiHooks?.onPlayerEffect(source, targets, effect.value.effect_type, [Number(effect.value.amount)]);
-                        player_block += effect.value.amount;
-                        break;
-                }
-            }
-        };
-
-        const targets = abilities.map((_, i) => aliveTargets[randIntBetween(rng, i, 0, aliveTargets.length - 1)]);
-        
-        // base effects
-        const allEnemiesDead = () => {
-            logger.combat.debug(`[${uiHooks == undefined}] checking damage: ${player_damage} | blocks: ${battleConfig.stats[0].block}, ${battleConfig.stats[1].block}, ${battleConfig.stats[2].block}   |  hp: ${battleState.enemy_hp_0}, ${battleState.enemy_hp_1}, ${battleState.enemy_hp_2}`);
-            return (player_damage[0] > battleConfig.stats[0].block + battleState.enemy_hp_0)
-                                  && (player_damage[1] > battleConfig.stats[1].block + battleState.enemy_hp_1 || enemy_count < 2)
-                                  && (player_damage[2] > battleConfig.stats[2].block + battleState.enemy_hp_2 || enemy_count < 3);
-        };
-        for (let i = 0; i < abilities.length; ++i) {
-            const ability = abilities[i];
-            await uiHooks?.onUseAbility(i, undefined);
-            if (ability.generate_color.is_some) {
-                await uiHooks?.onEnergyTrigger(i, Number(ability.generate_color.value));
-            }
-            await resolveEffect(ability.effect, i, targets[i]);
-            await uiHooks?.afterUseAbility(i);
-            if (allEnemiesDead()) {
-                logger.combat.debug(`[${uiHooks == undefined}] prematurely ending`);
-                return handleEndOfRound();
-            }
-        }
-        // energy triggers
-        for (let i = 0; i < abilities.length; ++i) {
-            const ability = abilities[i];
-            for (let c = 0; c < 3; ++c) {
-                if (ability.on_energy[c].is_some && abilities.some((a2, j) => i != j && a2.generate_color.is_some && Number(a2.generate_color.value) == c)) {
-                    await uiHooks?.onUseAbility(i, c);
-                    await resolveEffect(ability.on_energy[c], i, targets[i]);
-                    await uiHooks?.afterUseAbility(i);
-                    if (allEnemiesDead()) {
-                        return handleEndOfRound();
-                    }
-                }
-            }
-        }
-
-
-
-        // enemy damage
-        for (let i = 0; i < battleConfig.enemy_count; ++i) {
-            if (enemy_hp[i] > 0) {
-                const damage = battleConfig.stats[i].attack;
-                 enemy_damage += damage;
-                if (Number(damage) != 0) {
-                    await uiHooks?.onEnemyAttack(i, Number(damage));
-                }
-            }
-        }
-
-        handleEndOfRound();
-    });
-}
-
-/**
- * Combat logic that accepts specific player targets instead of using random targeting.
- * This is identical to combat_round_logic except it uses playerTargets instead of random selection.
- */
-export function combat_round_logic_with_targets(battle_id: bigint, gameState: Game2DerivedState, playerTargets: number[], uiHooks?: CombatCallbacks): Promise<BattleRewards | undefined> {
-    return new Promise(async (resolve) => {
-        logger.combat.debug(`combat_round_logic_with_targets(${playerTargets}, ${uiHooks == undefined})`);
-        const battleConfig = gameState.activeBattleConfigs.get(battle_id)!;
-        const battleState = gameState.activeBattleStates.get(battle_id)!;
-        const rng = pureCircuits.fakeTempRng(battleState, battleConfig);
         if (uiHooks != undefined) {
             gameState.ui = true;
         } else {
